SERVER-28657 Resolve missing edges and repair deps
A number of edges can now be resolved.  A few libraries were
made fully resolved by breaking the few TUs that depend upon
cycles into their own sub-libraries.  This should help prevent
the appearance of hidden cycles, as we progress towards a
cycle-free graph.

One library was missed as a cyclical-dependant of catalog in the
graph rotation of `catalog` to resolved.  This library was cut
from `catalog` to present remainging hidden cycles into `catalog`.