SERVER-27244 Status usage compile-time checking.
There are numerous places in the codebase where `mongo::Status` or
`mongo::StatusWith< T >` objects are returned and never checked.
Many of these are innocuous, but many of them are potentially
severe bugs.  This change introduces compile-time checking of
`Status` and `StatusWith` usage.  It also introduces an `ignore`
function which is useful to state that a specific "ignored status"
case was intentional.  It is now an error, in clang builds, to
forget to check a `Status`.

It also introduces a `transitional_ignore` function, which allows
for easy continual auditing of the codebase for current
unchecked-status instances.  All present "ignored status" cases
have been marked `transitional_ignore`.