SERVER-36967 Expanded static checking on status.
Unused status objects and unchecked return codes from status objects
are now flagged by the compiler.  This has found a number of cases
where the ball was dropped on checking status results.  Further,
unit test machinery for status-returning functions in contexts which
check for exceptions that are thrown has been relaxed to not require
a manual ignore statement.  We should still consider cutting down on
the number of status-returning-and-throwing functions, but that is
beyond the scope of this change.